solution/requirements.txt
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==5.0.0
billiard==4.2.4
black==23.11.0
blinker==1.9.0
celery==5.6.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.120.2
fastapi-cache2==0.2.2
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastar==0.8.0
flake8==6.1.0
flower==2.0.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.14.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
MyApplication==0.1.0
mypy==1.7.1
mypy_extensions==1.1.0
openpyxl==3.1.2
orjson==3.11.4
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.1.0
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==4.0.0
python-multipart==0.0.20
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==4.6.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.2
routers==0.10.1
rsa==4.9.1
sentry-sdk==2.43.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
tornado==6.5.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
uliweb-alembic==0.6.9
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
